+++ package.json
@@ -5,8 +5,7 @@
   "type": "module",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node script",
-    "": "dfqad"
+    "start": "node script"
   },
   "repository": {
     "type": "git",
